+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Allow Bitcoin fee rates with 0.01 precision
 - Reserve 1200 sats of LBTC balance to save on Elements fees
 - Show confidential peg-in join time limit on peg-in form
+- Allow updating txid of externally funded peg-in when not found
 
 ## 1.7.6